+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Added a timestamp to the appcache file for better cache versioning ([#4][])
 * Remove use of deprecated State mixin ([#5][])
 * Push states in the browser history when changing day or track ([#6][])
+* Encode talk titles ([#7][])
 * Code updated to use React 0.13.x
 * Dependencies updated
 
@@ -13,6 +14,7 @@
 [#4]: https://github.com/AurelioDeRosa/ConfAgenda/issues/4
 [#5]: https://github.com/AurelioDeRosa/ConfAgenda/issues/5
 [#6]: https://github.com/AurelioDeRosa/ConfAgenda/issues/6
+[#7]: https://github.com/AurelioDeRosa/ConfAgenda/issues/7
 
 # 0.1.0 (2015-02-10) #